Fixed warnings

+++ miocpp/FileReader.cpp
@@ -1,10 +1,11 @@
 #include "FileReader.h"
+#include <stdlib.h>
 
 FileReader::~FileReader() {
 }
 
 void FileReader::readFully(char* buffer, size_t length) {
-    size_t red;
+    int red;
     for (size_t pos = 0; pos < length; pos += red) {
         red = read(buffer+pos, length - pos);
         if (red < 0) {
